Created the equation for scale alignment!
I was discouraged working on this in class the other day, and my friend Logan said "nah man, you'll get it" and I really didn't believe him. but I sat down for 1 more hour and did some more investigation into the scaling function and man, let's go I figured out the issue with the getrotationalmatrix as I had copied the code from stack overflow, it was so simple now that I understand. 
I also checked my code with the webcam so that I coiuld see what would happen as I made my face closer and further back from the camera. going to align some photos now and reflect the new gif on my website noahbuchanan.me! It's cool because the scale inadvertendly aligns the right eye as well. An unintended feature but a good choice on my end for choosing to align the left eye not the part between the eyes.